+++ src/file_dialog.rs
@@ -178,6 +178,12 @@ impl Debug for dyn FileDialogModal + Send + Sync {
     }
 }
 
+/// Callback type to inject a custom egui ui inside the file dialog's ui.
+///
+/// Also gives access to the file dialog, since it would otherwise be inaccessible
+/// inside the closure.
+type FileDialogUiCallback<'a> = dyn FnMut(&mut egui::Ui, &mut FileDialog) + 'a;
+
 impl FileDialog {
     // ------------------------------------------------------------------------
     // Creation:
@@ -380,7 +386,33 @@ impl FileDialog {
         }
 
         self.update_keybindings(ctx);
-        self.update_ui(ctx);
+        self.update_ui(ctx, None);
+
+        self
+    }
+
+    /// Do an [update](`Self::update`) with a custom right panel ui.
+    ///
+    /// Example use cases:
+    /// - Show custom information for a file (size, MIME type, etc.)
+    /// - Embed a preview, like a thumbnail for an image
+    /// - Add controls for custom open options, like open as read-only, etc.
+    ///
+    /// See [`active_entry`](Self::active_entry) to get the active directory entry
+    /// to show the information for.
+    ///
+    /// This function has no effect if the dialog state is currently not `DialogState::Open`.
+    pub fn update_with_right_panel_ui(
+        &mut self,
+        ctx: &egui::Context,
+        f: &mut FileDialogUiCallback,
+    ) -> &Self {
+        if self.state != DialogState::Open {
+            return self;
+        }
+
+        self.update_keybindings(ctx);
+        self.update_ui(ctx, Some(f));
 
         self
     }
@@ -859,6 +891,16 @@ impl FileDialog {
         self
     }
 
+    /// Sets whether the show system files option inside the top panel
+    /// menu should be visible.
+    ///
+    /// Has no effect when `FileDialog::show_top_panel` or
+    /// `FileDialog::show_menu_button` is disabled.
+    pub const fn show_system_files_option(mut self, show_system_files_option: bool) -> Self {
+        self.config.show_system_files_option = show_system_files_option;
+        self
+    }
+
     /// Sets whether the search input should be visible in the top panel.
     ///
     /// Has no effect when `FileDialog::show_top_panel` is disabled.
@@ -969,6 +1011,26 @@ impl FileDialog {
         }
     }
 
+    /// Returns the currently active directory entry.
+    ///
+    /// This is either the currently highlighted entry, or the currently active directory
+    /// if nothing is being highlighted.
+    ///
+    /// For the [`DialogMode::SelectMultiple`] counterpart,
+    /// see [`FileDialog::active_selected_entries`].
+    pub const fn active_entry(&self) -> Option<&DirectoryEntry> {
+        self.selected_item.as_ref()
+    }
+
+    /// Returns an iterator over the currently selected entries in [`SelectMultiple`] mode.
+    ///
+    /// For the counterpart in single selection modes, see [`FileDialog::active_entry`].
+    ///
+    /// [`SelectMultiple`]: DialogMode::SelectMultiple
+    pub fn active_selected_entries(&self) -> impl Iterator<Item = &DirectoryEntry> {
+        self.get_dir_content_filtered_iter().filter(|p| p.selected)
+    }
+
     /// Returns the ID of the operation for which the dialog is currently being used.
     ///
     /// See `FileDialog::open` for more information.
@@ -990,7 +1052,13 @@ impl FileDialog {
 /// UI methods
 impl FileDialog {
     /// Main update method of the UI
-    fn update_ui(&mut self, ctx: &egui::Context) {
+    ///
+    /// Takes an optional callback to show a custom right panel.
+    fn update_ui(
+        &mut self,
+        ctx: &egui::Context,
+        right_panel_fn: Option<&mut FileDialogUiCallback>,
+    ) {
         let mut is_open = true;
 
         if self.config.as_modal {
@@ -1019,6 +1087,17 @@ impl FileDialog {
                     .width_range(90.0..=250.0)
                     .show_inside(ui, |ui| {
                         self.ui_update_left_panel(ui);
+                    });
+            }
+
+            // Optionally, show a custom right panel (see `update_with_custom_right_panel`)
+            if let Some(f) = right_panel_fn {
+                egui::SidePanel::right(self.window_id.with("right_panel"))
+                    // Unlike the left panel, we have no control over the contents, so
+                    // we don't restrict the width. It's up to the user to make the UI presentable.
+                    .resizable(true)
+                    .show_inside(ui, |ui| {
+                        f(ui, self);
                     });
             }
 
@@ -1166,7 +1245,9 @@ impl FileDialog {
 
             // Menu button containing reload button and different options
             if self.config.show_menu_button
-                && (self.config.show_reload_button || self.config.show_hidden_option)
+                && (self.config.show_reload_button
+                    || self.config.show_hidden_option
+                    || self.config.show_system_files_option)
             {
                 ui.allocate_ui_with_layout(
                     BUTTON_SIZE,
@@ -1185,6 +1266,18 @@ impl FileDialog {
                                     .checkbox(
                                         &mut self.config.storage.show_hidden,
                                         &self.config.labels.show_hidden,
+                                    )
+                                    .clicked()
+                            {
+                                self.refresh();
+                                ui.close_menu();
+                            }
+
+                            if self.config.show_system_files_option
+                                && ui
+                                    .checkbox(
+                                        &mut self.config.storage.show_system_files,
+                                        &self.config.labels.show_system_files,
                                     )
                                     .clicked()
                             {
@@ -2494,8 +2587,7 @@ impl FileDialog {
             }
             DialogMode::SelectMultiple => {
                 let result: Vec<PathBuf> = self
-                    .get_dir_content_filtered_iter()
-                    .filter(|p| p.selected)
+                    .active_selected_entries()
                     .map(crate::DirectoryEntry::to_path_buf)
                     .collect();
 
@@ -2558,11 +2650,10 @@ impl FileDialog {
     /// What is checked depends on the mode the dialog is currently in.
     fn is_selection_valid(&self) -> bool {
         match &self.mode {
-            DialogMode::SelectDirectory => {
-                self.selected_item
-                    .as_ref()
-                    .map_or(false, crate::DirectoryEntry::is_dir)
-            }
+            DialogMode::SelectDirectory => self
+                .selected_item
+                .as_ref()
+                .map_or(false, crate::DirectoryEntry::is_dir),
             DialogMode::SelectFile => self
                 .selected_item
                 .as_ref()
@@ -2863,6 +2954,7 @@ impl FileDialog {
             path,
             self.show_files,
             self.config.storage.show_hidden,
+            self.config.storage.show_system_files,
             self.get_selected_file_filter(),
         );